030_LLM_Interface_test/pyproject.toml
[project]
name = "030-llm-interface-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "gradio>=5.35.0",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
    "streamlit>=1.46.1",
]
